keys: move autogen function to core
